% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denscomp.R
\name{denscomp}
\alias{denscomp}
\title{Components Density Plot}
\usage{
denscomp(
  model,
  sep = "+",
  combination,
  violin = FALSE,
  random = TRUE,
  z_value = TRUE
)
}
\arguments{
\item{model}{An object of class \code{\link[netmeta]{netmeta}}.}

\item{sep}{A single character that defines the separator between interventions components.}

\item{combination}{A character vector that contains the component combinations of interest.}

\item{violin}{\code{logical}. If \code{TRUE} the density is visualized via violins instead of density plots.}

\item{random}{\code{logical}. If \code{TRUE} the random-effects NMA model is used, instead of the fixed-effect NMA model.}

\item{z_value}{\code{logical}. If \code{TRUE} z-values are used, instead intervention effects.}
}
\value{
An object of class \code{ggplot}.
}
\description{
The function creates density plots in order to explore the efficacy of the components.
}
\details{
If the length of the argument \code{combination} is 1, the function creates two density plots. The first is produced based on the
interventions that include the component combination of interest (which is specified by the argument \code{combination}),
while the second on the interventions that do not include the underlying component combination.

If the argument \code{combination} includes more than one elements, the number of densities is equal with the length of
the argument \code{combination}, and each density is based on the interventions that include the relative component combination.
For example, if \code{combination = c("A + B", "B + C", "A")} the function will produce 3 density plots that are based on
the interventions that includes components \code{"A"} and \code{"B"}, the interventions that include components \code{"B"} and \code{"C"} and
interventions that includes component \code{"A"}, respectively.

The function by default uses the intervention's z-values \code{z_value = TRUE} obtained from the random-effects network
meta-analysis (NMA) model (\code{random = TRUE}). It could be also adjusted to use the intervention's effect estimates
instead of the z-values, by setting \code{z_value = FALSE}.
}
\note{
The efficacy of the components could be explored via violins plots instead of density plots, by setting \code{violin = TRUE}.
}
\examples{
data(MACE)
NMAdata <- netmeta::pairwise(
  studlab = Study, treat = list(treat1, treat2, treat3, treat4),
  n = list(n1, n2, n3, n4), event = list(event1, event2, event3, event4), data = MACE, sm = "OR"
)
net <- netmeta::netmeta(
  TE = TE, seTE = seTE, studlab = studlab, treat1 = treat1,
  treat2 = treat2, data = NMAdata, ref = "UC"
)
denscomp(model = net, combination = "C")
}
