% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACE.R
\docType{data}
\name{MACE}
\alias{MACE}
\title{Major Adverse Cardiovascular Event}
\format{
A \code{data.frame} with the following columns
\describe{
\item{Study}{The study name of the trial}
\item{treat1, treat2, treat3, treat4}{Treatment names of arms}
\item{n1, n2, n3, n4}{Total number of participants in arms}
\item{event1, event2, event3, event4}{Total number of events in arms}
}
}
\usage{
MACE
}
\description{
An artificial network meta-analysis data compering the effectiveness of a number of interventions for major
adverse cardiovascular events.
}
\keyword{datasets}
