% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmaMACE.R
\docType{data}
\name{nmaMACE}
\alias{nmaMACE}
\title{Network Meta-Analysis of Major Adverse Cardiovascular Event}
\format{
An object of class \code{netmeta} of length 157.
}
\usage{
nmaMACE
}
\description{
An artificial network meta-analysis (of class \code{\link[netmeta]{netmeta}}) compering the effectiveness of a number of interventions for major
adverse cardiovascular events.
}
\keyword{datasets}
