% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatcomp.R
\name{heatcomp}
\alias{heatcomp}
\title{Components Heat Plot}
\usage{
heatcomp(
  model,
  sep = "+",
  median = TRUE,
  random = TRUE,
  z_value = TRUE,
  freq = TRUE,
  outcome = "beneficial",
  legend_name = NULL
)
}
\arguments{
\item{model}{An object of class \code{\link[netmeta]{netmeta}}.}

\item{sep}{A single character that defines the separator between interventions components.}

\item{median}{\code{logical}. If \code{TRUE} the median is used instead of the mean as a summary measure.}

\item{random}{\code{logical}. If \code{TRUE} the random-effects NMA model is used instead of the fixed-effect NMA model.}

\item{z_value}{\code{logical}. If \code{TRUE} z-values are used instead of interventions effects.}

\item{freq}{\code{logical}. If \code{TRUE} the frequency of component combinations are printed.}

\item{outcome}{A single character that defines the outcome's nature. Either \code{beneficial}, or \code{harmful} can be abbreviated.}

\item{legend_name}{A single character that specifies the title of the legend.}
}
\value{
An object of class \code{ggplot}.
}
\description{
The function creates a heat plot based on the two-by-two component combinations, obtained from the
network meta-analysis (NMA) model.
}
\details{
Diagonal elements refer to components, while off-diagonal to components' combinations. Each element summarizes
the z-values (if \code{z_value = TRUE}) of the interventions that includes the corresponding
component combination. Z-scores quantify the strength of statistical evidence. Combinations that were not
observed in the NMA model, are denoted by the letter "X". Frequency of component combinations observed in the NMA is printed
by default (\code{freq = TRUE}). As a summary measure, the median is used by default (\code{median = TRUE}).
The magnitude of each z-score is reflected by the color's intensity. Estimates close to zero
are denoted by white color, and indicates lack of evidence that the corresponding component combination performs
better than the reference intervention. Deep green and red colors indicate strong evidence that the corresponding
component combination performs better and worse respectively. Argument \code{outcome} controls the
outcomes nature (beneficial or harmful). For harmful outcomes small values are considered as good whereas the
opposite holds for beneficial outcomes.

Intervention effects can be used instead of z-values by setting \code{z_value = FALSE}.
Also, by setting \code{median = FALSE}, the mean is used instead of the median as a summary measure.
}
\note{
In the case of dichotomous outcomes, the log-scale is used.
}
\examples{
data(MACE)
NMAdata <- netmeta::pairwise(
  studlab = Study, treat = list(treat1, treat2, treat3, treat4),
  n = list(n1, n2, n3, n4), event = list(event1, event2, event3, event4), data = MACE, sm = "OR"
)
net <- netmeta::netmeta(
  TE = TE, seTE = seTE, studlab = studlab, treat1 = treat1,
  treat2 = treat2, data = NMAdata, ref = "UC"
)
heatcomp(model = net)
}
